docs: fix links to WebAuthn specification
Most links to the Web Authentication API specification are broken in the
sense that the anchor IDs do not exist anymore. This is because the current
"base" URL of these links points to the level 2 specification. The result is
that there is no auto-scroll to the sections identified by the anchor IDs.
This commit fixes that by making all links explicitly point to the
level 1 specification.